requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.2
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.40.3
idna==3.10
oauthlib==3.3.1
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
